pyproject.toml
[project]
name = "survey-of-chatbot-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "huggingface-hub>=0.36.0",
    "transformers>=4.48.3",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "torchaudio==2.9.1",
    "llama-index>=0.14.13",
    "llama-index-llms-huggingface>=0.6.1",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-postprocessor-flag-embedding-reranker>=0.4.1",
    "flagembedding",
    "raganything[all]",
    "lightrag-hku",
    "sentence-transformers>=5.2.2",
    "typer>=0.21.1",
    "hf-xet>=1.2.0",
    "flashrank>=0.2.10",
    "faiss-cpu>=1.13.2",
    "pymupdf>=1.27.1",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
flagembedding = { git = "https://github.com/FlagOpen/FlagEmbedding.git" }
raganything = { git = "https://github.com/HKUDS/RAG-Anything.git" }
lightrag-hku = { git = "https://github.com/HKUDS/LightRAG.git" }
